Read and display the smallest temperature across all cities

+++ app/src/main/java/pakiet/arkadiuszzimny/recruitmentweatherapp/HourlyTemp.kt
@@ -2,5 +2,5 @@ package pakiet.arkadiuszzimny.recruitmentweatherapp
 
 data class HourlyTemp(
     val hour: Int,
-    val temp: Int
+    val temp: Double
 )
